+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.19.0] - 2021-04-16
+
 ## Added
 
 - The voter `encrypt` method now returns the `plainVote` and the `electionUniqueId` so it will be included when the user audits the ballot.